Work around np.nan not having shape (while other numpy floats do)

+++ hyperspy/_signals/signal1d.py
@@ -223,7 +223,7 @@ def _estimate_shift1D(data, **kwargs):
     ip = kwargs.get('ip', 5)
     data_slice = kwargs.get('data_slice', slice(None))
     if bool(mask):
-        return np.nan
+        return np.float32(np.nan)
     data = data[data_slice]
     if interpolate is True:
         data = interpolate1D(ip, data)
